refactor(components): optimize component design
