+++ app/src/main/java/nextstep/shoppingcart/presentation/components/topbars/TopBarWithNavigation.kt
@@ -1,0 +1,57 @@
+package nextstep.shoppingcart.presentation.components.topbars
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TopBarWithNavigation(
+    name: String = "상품 상세",
+    navigation: () -> Unit = {},
+    modifier: Modifier = Modifier
+) {
+    TopAppBar(
+        title = {
+            Text(
+                text = name,
+                style =
+                TextStyle(
+                    fontSize = 22.sp,
+                    fontWeight = FontWeight.Bold
+                )
+            )
+        },
+        modifier = modifier,
+        navigationIcon = {
+            IconButton(
+                onClick = { navigation() }
+            ) {
+                Icon(
+                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                    contentDescription = "Back"
+                )
+            }
+        }
+    )
+}
+
+
+@Preview
+@Composable
+private fun BarLayoutPreview() {
+    Column {
+        TopBarWithNavigation(name = "상품 상세")
+    }
+}
