+++ app/src/main/java/nextstep/shoppingcart/ui/theme/Color.kt
@@ -9,3 +9,5 @@ val Pink80 = Color(0xFFEFB8C8)
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)
+
+val Blue50 = Color(0xFF2196F3)
